Backend/requirements.txt
Flask==2.3.2
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-oauthlib==0.8.0
python-dateutil==2.8.2
